Auto-throttling and auto-blacklisting!
There are now three layers of automatic abuse prevention. The old manual
blacklist has been retired.

A request with a naughtiness score >= 0.5 will be throttled, making things
slow for the client.

A request with a score >= 1 will be blacklisted, returning an error.

A request with a score >= 3 will receive evil CSS or JS.