Improve how upstream errors are handled.
GitHub is responsible for most errors people receive from RawGit, but we
haven't done a good job of making that clear. Now, instead of treating
all upstream errors as 404s, we pass through the upstream error codes
and responses.

If we can't connect to GitHub or if GitHub sends an invalid response, we
return a 502 error that clearly explains what happened.